requirements.txt
pygame
neat-python
argparse